+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to Kimia will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.23] - 2026-02-13
+## [Unreleased]
 ### Added
 - Added --digest-file export handling for BuildKit output.