+++ CHANGELOG.md
@@ -5,8 +5,7 @@ All notable changes to Kimia will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
-## Unreleased
+## [Unreleased]
 
 ### Added
 
@@ -14,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix storage drivers for BuildKit and Buildah
+- Remediations and updates for security scanner findings
+
+### Removed
 
 ## [1.0.22] - 2025-12-02