+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added --digest-file export handling for BuildKit output.
+- Added --reproducible flag support for BuildKit provenance attestations.
 
 ### Changed